update MenuBar + format Explorer for better readability + fix overflow in Explorer
