+++ src/scss/_variables.scss
@@ -7,3 +7,8 @@ $taskBarHoverBg: #122830f3;
 
 $titleBarHeight: 40px;
 $titleBarBg: white;
+
+$itemHoverBg: #d5ecfc;
+$itemFocusBg: #cce8fe;
+$itemBorderColor: #8ecbfa;
+$menuBarBorder: #f1f1f1;
