+++ src/softwares/Explorer/components/Icon/Icon.jsx
@@ -1,0 +1,11 @@
+import { FiHardDrive } from 'react-icons/fi';
+import { FcDocument } from 'react-icons/fc';
+
+const Icon = ({ isDir }) =>
+  isDir ? (
+    <img alt="Folder" src={`${process.env.PUBLIC_URL}/icons/MiFolder.svg`} />
+  ) : (
+    <FcDocument fill="white" />
+  );
+
+export default Icon;
