+++ src/softwares/Explorer/components/Ribbon/Ribbon.scss
