+++ src/softwares/Explorer/components/Sidebar/Sidebar.scss
